pyproject.toml
[tool.poetry]
name = "kulakoff1988_gendiff"
version = "0.1.3"
description = "The program for files comparison"
authors = ["Aleksandr Kulakov <kulakofft4@gmail.com>"]
repository = "https://test.pypi.org/project/kulakoff1988-brain-gendiff/"
packages = [
	{ include = "gendiff" }
]
classifiers = [
	"Programming Language :: Python :: 3",
	"Operating System :: OS Independent",
	"Environment :: Console",
	"Development Status :: 5 - Production/Stable",
	"Intended Audience :: End Users/Desktop",
	"License :: Freeware",
	"Natural Language :: English"
]

[tool.poetry.dependencies]
python = "^3.6"
PyYAML = "^5.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"
pytest = "^5.0"
coverage = "^4.5"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
